Update forms to include dropdown
Updates the create and update plate appearance forms to include
dropdowns for options where users need to select between several
options.

Also includes a new helper called 'noeq' that tests for inequality
in handlebars if statements, which was used to pass in the current
values into the edit form dropdowns without repeating it in the
options.