Dynamic Box Score started
Started a dynamic box score feature that updates automatically as a
user adds plate appearances to the game.

Commit contains files for the box score component and logic added
to the game and PA-edit routes to calculate the score and update
the box score whenever a PA is created, destroyed or updated.