Addons and extra inning functionality
Adds addons for filter/sorting for the games table and a date picker
that will work in all browsers.

Also expands scorekeeping functionality to work with extra innings and
dynamically display extra innings on scoreboard as they are added
via plate appearances. Now supports games that go as high as 18
innings, with additional support for 18+ innings coming later.